artist pagination update
